Combine some files and use @import jsdoc
