+++ src/datapage.js
@@ -1,3 +1,4 @@
+import { deltaBinaryUnpack, deltaByteArray, deltaLengthByteArray } from './delta.js'
 import { bitWidth, byteStreamSplit, readRleBitPackedHybrid } from './encoding.js'
 import { readPlain } from './plain.js'
 import { getMaxDefinitionLevel, getMaxRepetitionLevel } from './schema.js'
@@ -6,7 +7,6 @@ import { snappyUncompress } from './snappy.js'
 /**
  * Read a data page from uncompressed reader.
  *
- * @import {ColumnMetaData, CompressionCodec, Compressors, DataPage, DataPageHeader, DataReader, DecodedArray, DictionaryPageHeader, SchemaTree} from '../src/types.js'
  * @param {Uint8Array} bytes raw page data (should already be decompressed)
  * @param {DataPageHeader} daph data page header
  * @param {SchemaTree[]} schemaPath
@@ -66,6 +66,7 @@ export function readDictionaryPage(bytes, diph, columnMetadata, typeLength) {
 }
 
 /**
+ * @import {ColumnMetaData, CompressionCodec, Compressors, DataPage, DataPageHeader, DataPageHeaderV2, DataReader, DecodedArray, DictionaryPageHeader, PageHeader, SchemaTree} from '../src/types.d.ts'
  * @param {DataReader} reader data view for the page
  * @param {DataPageHeader} daph data page header
  * @param {SchemaTree[]} schemaPath
@@ -131,4 +132,111 @@ export function decompressPage(compressedBytes, uncompressed_page_size, codec, c
     throw new Error(`parquet decompressed page length ${page?.length} does not match header ${uncompressed_page_size}`)
   }
   return page
+}
+
+
+/**
+ * Read a data page from the given Uint8Array.
+ *
+ * @param {Uint8Array} compressedBytes raw page data
+ * @param {PageHeader} ph page header
+ * @param {SchemaTree[]} schemaPath
+ * @param {ColumnMetaData} columnMetadata
+ * @param {Compressors | undefined} compressors
+ * @returns {DataPage} definition levels, repetition levels, and array of values
+ */
+export function readDataPageV2(compressedBytes, ph, schemaPath, columnMetadata, compressors) {
+  const view = new DataView(compressedBytes.buffer, compressedBytes.byteOffset, compressedBytes.byteLength)
+  const reader = { view, offset: 0 }
+  const { codec, type } = columnMetadata
+  const daph2 = ph.data_page_header_v2
+  if (!daph2) throw new Error('parquet data page header v2 is undefined')
+
+  // repetition levels
+  const repetitionLevels = readRepetitionLevelsV2(reader, daph2, schemaPath)
+  reader.offset = daph2.repetition_levels_byte_length // readVarInt() => len for boolean v2?
+
+  // definition levels
+  const definitionLevels = readDefinitionLevelsV2(reader, daph2, schemaPath)
+  // assert(reader.offset === daph2.repetition_levels_byte_length + daph2.definition_levels_byte_length)
+
+  const uncompressedPageSize = ph.uncompressed_page_size - daph2.definition_levels_byte_length - daph2.repetition_levels_byte_length
+
+  let page = compressedBytes.subarray(reader.offset)
+  if (daph2.is_compressed !== false) {
+    page = decompressPage(page, uncompressedPageSize, codec, compressors)
+  }
+  const pageView = new DataView(page.buffer, page.byteOffset, page.byteLength)
+  const pageReader = { view: pageView, offset: 0 }
+
+  // read values based on encoding
+  /** @type {DecodedArray} */
+  let dataPage
+  const nValues = daph2.num_values - daph2.num_nulls
+  if (daph2.encoding === 'PLAIN') {
+    const { type_length } = schemaPath[schemaPath.length - 1].element
+    dataPage = readPlain(pageReader, type, nValues, type_length)
+  } else if (daph2.encoding === 'RLE') {
+    // assert(columnMetadata.type === 'BOOLEAN')
+    dataPage = new Array(nValues)
+    readRleBitPackedHybrid(pageReader, 1, 0, dataPage)
+    dataPage = dataPage.map(x => !!x)
+  } else if (
+    daph2.encoding === 'PLAIN_DICTIONARY' ||
+    daph2.encoding === 'RLE_DICTIONARY'
+  ) {
+    const bitWidth = pageView.getUint8(pageReader.offset++)
+    dataPage = new Array(nValues)
+    readRleBitPackedHybrid(pageReader, bitWidth, uncompressedPageSize - 1, dataPage)
+  } else if (daph2.encoding === 'DELTA_BINARY_PACKED') {
+    const int32 = type === 'INT32'
+    dataPage = int32 ? new Int32Array(nValues) : new BigInt64Array(nValues)
+    deltaBinaryUnpack(pageReader, nValues, dataPage)
+  } else if (daph2.encoding === 'DELTA_LENGTH_BYTE_ARRAY') {
+    dataPage = new Array(nValues)
+    deltaLengthByteArray(pageReader, nValues, dataPage)
+  } else if (daph2.encoding === 'DELTA_BYTE_ARRAY') {
+    dataPage = new Array(nValues)
+    deltaByteArray(pageReader, nValues, dataPage)
+  } else if (daph2.encoding === 'BYTE_STREAM_SPLIT') {
+    const { type_length } = schemaPath[schemaPath.length - 1].element
+    dataPage = byteStreamSplit(reader, nValues, type, type_length)
+  } else {
+    throw new Error(`parquet unsupported encoding: ${daph2.encoding}`)
+  }
+
+  return { definitionLevels, repetitionLevels, dataPage }
+}
+
+/**
+ * @param {DataReader} reader
+ * @param {DataPageHeaderV2} daph2 data page header v2
+ * @param {SchemaTree[]} schemaPath
+ * @returns {any[]} repetition levels
+ */
+function readRepetitionLevelsV2(reader, daph2, schemaPath) {
+  const maxRepetitionLevel = getMaxRepetitionLevel(schemaPath)
+  if (!maxRepetitionLevel) return []
+
+  const values = new Array(daph2.num_values)
+  readRleBitPackedHybrid(
+    reader, bitWidth(maxRepetitionLevel), daph2.repetition_levels_byte_length, values
+  )
+  return values
+}
+
+/**
+ * @param {DataReader} reader
+ * @param {DataPageHeaderV2} daph2 data page header v2
+ * @param {SchemaTree[]} schemaPath
+ * @returns {number[] | undefined} definition levels
+ */
+function readDefinitionLevelsV2(reader, daph2, schemaPath) {
+  const maxDefinitionLevel = getMaxDefinitionLevel(schemaPath)
+  if (maxDefinitionLevel) {
+    // V2 we know the length
+    const values = new Array(daph2.num_values)
+    readRleBitPackedHybrid(reader, bitWidth(maxDefinitionLevel), daph2.definition_levels_byte_length, values)
+    return values
+  }
 }
